Add GCC build preset and fix non-vendor build errors for GCC.

+++ src/Client/techniques/shadows/CascadedShadowMaps.cpp
@@ -94,7 +94,7 @@ void Techniques::CascadedShadowMap::RenderTerrainShadowMap(Scheduler& scheduler,
 
       for (uint32_t i = 0; i < params.numCascades; i++)
       {
-        const auto side                    = (params.baseFrustumSideLength / 2) * exp2(float(i));
+        const auto side                    = static_cast<float>((params.baseFrustumSideLength / 2) * exp2(float(i)));
         const auto clip_from_view          = glm::ortho(-side, side, -side, side, -params.frustumDepth / 2, params.frustumDepth / 2);
         shadow.cascades[i].clip_from_world = clip_from_view * view_from_world;
         shadow.cascades[i].world_from_clip = glm::inverse(shadow.cascades[i].clip_from_world);
